Moves pillarProps & sanitizeOptions to utils folder
