+++ src/utils/sanitizeOptions.ts
